refactor(constants): extract configuration data to dedicated files
Split the 271-line constants.dart into 9 logical sub-files under
lib/constants/. The original file becomes a barrel re-export so all
23 callers continue to work without import changes.

No behavior changes, no new dependencies.
