+++ lib/constants/anime_tags.dart
@@ -1,0 +1,5 @@
+final List<String> defaultAnimeTags = const [
+  '日常', '原创', '校园', '搞笑', '奇幻',
+  '百合', '恋爱', '悬疑', '热血', '后宫',
+  '机战', '轻改', '偶像', '治愈', '异世界',
+];
