+++ lib/constants/decoders.dart
@@ -1,0 +1,36 @@
+/// 可选硬件解码器
+const Map<String, String> hardwareDecodersList = {
+  'auto': '启用任意可用解码器',
+  'auto-safe': '启用最佳解码器',
+  'auto-copy': '启用带拷贝功能的最佳解码器',
+  'd3d11va': 'DirectX11 (windows8 及以上)',
+  'd3d11va-copy': 'DirectX11 (windows8 及以上) (非直通)',
+  'videotoolbox': 'VideoToolbox (macOS / iOS)',
+  'videotoolbox-copy': 'VideoToolbox (macOS / iOS) (非直通)',
+  'vaapi': 'VAAPI (Linux)',
+  'vaapi-copy': 'VAAPI (Linux) (非直通)',
+  'nvdec': 'NVDEC (NVIDIA独占)',
+  'nvdec-copy': 'NVDEC (NVIDIA独占) (非直通)',
+  'drm': 'DRM (Linux)',
+  'drm-copy': 'DRM (Linux) (非直通)',
+  'vulkan': 'Vulkan (全平台) (实验性)',
+  'vulkan-copy': 'Vulkan (全平台) (实验性) (非直通)',
+  'dxva2': 'DXVA2 (Windows7 及以上)',
+  'dxva2-copy': 'DXVA2 (Windows7 及以上) (非直通)',
+  'vdpau': 'VDPAU (Linux)',
+  'vdpau-copy': 'VDPAU (Linux) (非直通)',
+  'mediacodec': 'MediaCodec (Android)',
+  'mediacodec-copy': 'MediaCodec (Android) (非直通)',
+  'cuda': 'CUDA (NVIDIA独占) (过时)',
+  'cuda-copy': 'CUDA (NVIDIA独占) (过时) (非直通)',
+  'crystalhd': 'CrystalHD (全平台) (过时)',
+  'rkmpp': 'Rockchip MPP (仅部分Rockchip芯片)',
+};
+
+/// Android 可选视频渲染器
+const Map<String, String> androidVideoRenderersList = {
+  'auto': '自动选择',
+  'gpu': '基于 OpenGL, 通用和稳健的选项',
+  'gpu-next': '基于 Vulkan, 在新设备上表现最好',
+  'mediacodec_embed': '功耗最低，不支持超分辨率',
+};
