+++ lib/constants/syncplay.dart
@@ -1,0 +1,10 @@
+/// 默认 SyncPlay 服务器列表
+const List<String> defaultSyncPlayEndPoints = [
+  'syncplay.pl:8995',
+  'syncplay.pl:8996',
+  'syncplay.pl:8997',
+  'syncplay.pl:8998',
+  'syncplay.pl:8999',
+];
+
+const String defaultSyncPlayEndPoint = 'syncplay.pl:8996';
